Remove automatic github_issue script and change instructions.
Getting the script to run wastes too much time on something that is not
core material, and all the actual instructions are still in the readme,
not in the issues, so we decided to just get rid of this step as it
doesn't really add that much value.

I added a slightly longer introduction to collaboration and the
suggestion to use issues and projects, with some links documenting
GitHubs automation features.